# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para sistema de inventario y ventas de ferretería.",
        "title": "Ferretería Inventario API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/api/products": {
            "get": {
                "description": "GET lista productos, POST crea producto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Listar o crear productos",
                "parameters": [
                    {
                        "description": "Producto (solo POST)",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Product"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Product"
                        }
                    }
                }
            },
            "post": {
                "description": "GET lista productos, POST crea producto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Listar o crear productos",
                "parameters": [
                    {
                        "description": "Producto (solo POST)",
                        "name": "product",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Product"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Product"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Product"
                        }
                    }
                }
            }
        },
        "/api/report/top-productos": {
            "get": {
                "description": "Devuelve los 5 productos más vendidos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Top productos vendidos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "additionalProperties": true
                            }
                        }
                    }
                }
            }
        },
        "/api/report/ventas-hoy": {
            "get": {
                "description": "Devuelve cantidad de ventas y total del día",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Ventas del día",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/sales": {
            "get": {
                "description": "GET lista ventas, POST crea venta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Listar o crear ventas",
                "parameters": [
                    {
                        "description": "Venta (solo POST)",
                        "name": "sale",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                        }
                    }
                }
            },
            "post": {
                "description": "GET lista ventas, POST crea venta",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Listar o crear ventas",
                "parameters": [
                    {
                        "description": "Venta (solo POST)",
                        "name": "sale",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                        }
                    }
                }
            }
        },
        "/api/sales/{id}": {
            "get": {
                "description": "Devuelve una venta con sus items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sales"
                ],
                "summary": "Obtener detalle de venta",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la venta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Sale"
                        }
                    }
                }
            }
        },
        "/clients": {
            "get": {
                "description": "GET lista clientes, POST crea cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Listar o crear clientes",
                "parameters": [
                    {
                        "description": "Cliente (solo POST)",
                        "name": "client",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Client"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Client"
                        }
                    }
                }
            },
            "post": {
                "description": "GET lista clientes, POST crea cliente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "Listar o crear clientes",
                "parameters": [
                    {
                        "description": "Cliente (solo POST)",
                        "name": "client",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Client"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Client"
                            }
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.Client"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ferreteria-inventario-ventas_internal_domain.Client": {
            "type": "object",
            "properties": {
                "cedula": {
                    "description": "Número de cédula (único)",
                    "type": "string"
                },
                "email": {
                    "description": "Correo electrónico",
                    "type": "string"
                },
                "id": {
                    "description": "Identificador único en la base de datos",
                    "type": "integer"
                },
                "nombre": {
                    "description": "Nombre completo del cliente",
                    "type": "string"
                }
            }
        },
        "ferreteria-inventario-ventas_internal_domain.Product": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Identificador único en la base de datos",
                    "type": "integer"
                },
                "nombre": {
                    "description": "Nombre del producto",
                    "type": "string"
                },
                "precio": {
                    "description": "Precio unitario del producto",
                    "type": "number"
                },
                "stock": {
                    "description": "Cantidad disponible en inventario",
                    "type": "integer"
                }
            }
        },
        "ferreteria-inventario-ventas_internal_domain.Sale": {
            "type": "object",
            "properties": {
                "client_id": {
                    "description": "Cliente que realiza la compra",
                    "type": "integer"
                },
                "fecha": {
                    "description": "Fecha de la venta",
                    "type": "string"
                },
                "id": {
                    "description": "ID de la venta",
                    "type": "integer"
                },
                "items": {
                    "description": "Lista de productos vendidos",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ferreteria-inventario-ventas_internal_domain.SaleItem"
                    }
                },
                "total": {
                    "description": "Total de la venta",
                    "type": "number"
                }
            }
        },
        "ferreteria-inventario-ventas_internal_domain.SaleItem": {
            "type": "object",
            "properties": {
                "cantidad": {
                    "description": "Cantidad vendida",
                    "type": "integer"
                },
                "precio_unitario": {
                    "description": "Precio al momento de la venta",
                    "type": "number"
                },
                "product_id": {
                    "description": "ID del producto vendido",
                    "type": "integer"
                },
                "subtotal": {
                    "description": "Cantidad * PrecioUnitario",
                    "type": "number"
                }
            }
        }
    }
}